Fix conflicting enum to_py function with multiple modules (#5887)
Fix conflicting names of cpdef enum to_py functions when the an
enum with the name name exists in multiple modules. Instead use
the cname to name the to_py function since we have already
ensured that it is unique and mangled with the module name.

Possibly fixes #5860 (it definitely fixes a real bug, but that
project has far too many dependencies for me to test, so who knows
if it fixes *that* bug).